Mirroring entire repo, instead of subset
Ran into an issue where installed packages on a machine would prevent an
update of packages required by the packages we want to install. For
example, docker requires a new verison of systemd, but existing packages
require the old one. This results in the upgrade of systemd being
blocked, because the existing package would have an unsatisfied
dependency.